% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data for week 7 in-class}
\format{
A data frame with 1028 rows and 4 variables:
\describe{
\item{y}{Binary Outcome}
\item{x1}{Continuous Predictor}
\item{x2}{Binary Predictor}
}
}
\source{
CHL 8010 course website
}
\usage{
toydata
}
\description{
Toy Dataset
}
\keyword{datasets}
