% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Calculate Odds Ratios with 95\% Confidence Intervals}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{Coefficients from a logistic regression model.}

\item{se}{Standard errors associated with the coefficients.}

\item{siglevel}{The significance level}

\item{roundto}{The precision for rounding the odds ratios and confidence intervals.}
}
\value{
A formatted string
}
\description{
This function computes the OR and its 95\% confidence intervals from a logistic regression model.
}
\examples{
OR_95CI(0.5, 0.1, 0.05, 2)
}
\author{
Mark Asuncion
}
